Porto: extra cell alignment

+++ Gainy/Gainy/Scenes/Portfolio/Plaid/Cells/BuyingPowerCell.swift
@@ -12,7 +12,7 @@ final class BuyingPowerCell: HoldingRangeableCell {
     
     @IBOutlet private weak var buyPowerLbl: UILabel!
     
-    var amount: Double = 0.0 {
+    var amount: Float = 0.0 {
         didSet {
             buyPowerLbl.text = amount.price
         }
